chromium/third_party/blink/web_tests/external/wpt/scroll-to-text-fragment/idlharness.window-expected.txt

This is a testharness.js-based test.
PASS idl_test setup
PASS idl_test validation
PASS Partial interface Location: original interface defined
PASS Partial interface Location: member names are unique
PASS FragmentDirective interface: existence and properties of interface object
PASS FragmentDirective interface object length
PASS FragmentDirective interface object name
PASS FragmentDirective interface: existence and properties of interface prototype object
PASS FragmentDirective interface: existence and properties of interface prototype object's "constructor" property
PASS FragmentDirective interface: existence and properties of interface prototype object's @@unscopables property
FAIL FragmentDirective must be primary interface of document.location.fragmentDirective assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of document.location.fragmentDirective assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Location interface: attribute fragmentDirective assert_true: The prototype object must have a property "fragmentDirective" expected true got false
FAIL Location interface: document.location must inherit property "fragmentDirective" with the proper type assert_inherits: property "fragmentDirective" not found in prototype chain
Harness: the test ran to completion.
